+++ main.go
@@ -1,12 +1,30 @@
 package main
 
 import (
-	log "github.com/sirupsen/logrus"
+	/*
+	"fmt"
+	"gitlab.com/ConsenSys/client/fr/core-stack/infra/key-store.git/secretstore"
+	"gitlab.com/ConsenSys/client/fr/core-stack/infra/key-store.git/keystore"
+	*/
 )
 
+// @title Swagger Example API
+// @version 1.0.1
 func main() {
-	var opts Config
-	LoadConfig(&opts)
-	ConfigureLogger(opts.Log)
-	log.Info("Start worker...")
+
+	/*
+	config := secretstore.VaultConfigFromViper()
+	hashicorpsSS, err := secretstore.NewHashicorps(config)
+
+	awsSS := secretstore.NewAWS(7)
+	tokenName := secretstore.VaultTokenFromViper()
+
+	err = hashicorpsSS.Init(awsSS, tokenName)
+	if err != nil {
+		fmt.Printf(err.Error())
+	}
+
+	keystore := keystore.NewBaseKeyStore(hashicorpsSS)
+	*/
+
 }