+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
-### Version 0.1.0
+### Version 0.2.0
 
-*Unreleased*
+*Released on March 22nd 2019*
+
+- [FEAT] Implement `KeyStore` interface
+- [FEAT] Implement Hashicorps `KeyStore`
+- [FEAT] Impelment configurations & flags
+- [FEAT] PreRegister private keys from configuration
+- [FEAT] Get Credentials from AWS